% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizExtremities.R
\name{vizExtremities}
\alias{vizExtremities}
\title{vizExtremities}
\usage{
vizExtremities(maxMemory = 5)
}
\arguments{
\item{maxMemory}{Maximum memory dedicated to the Shiny application (Go)}
}
\description{
Shiny application launcher
}
